{"schema":"decodex/commit/1","summary":"Move overlay window event runtime","authority":"manual"}
